requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
sentence-transformers
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
psycopg2
pinecone
pgvector